+++ feline/Matrix3d.h
@@ -3,30 +3,30 @@
 class Matrix3d
 {
 public:
-	float mat[3][3];
+	double mat[3][3];
 	Matrix3d(void);
-	Matrix3d(float[3][3]);
+	Matrix3d(double[3][3]);
 	Matrix3d(const Matrix3d& m);
-	Matrix3d(float a00,float a01,float a02,
-			 float a10,float a11,float a12,
-			 float a20,float a21,float a22);
+	Matrix3d(double a00,double a01,double a02,
+			 double a10,double a11,double a12,
+			 double a20,double a21,double a22);
 	inline Matrix3d transpose();
-	inline float determinant();
+	inline double determinant();
 	inline Matrix3d operator*(Matrix3d&);
-	inline Matrix3d operator*(float);
-	inline vector3<float> operator*(vector3<float>&);
+	inline Matrix3d operator*(double);
+	inline vector3<double> operator*(vector3<double>&);
 	inline Matrix3d operator+(Matrix3d&);
 	inline Matrix3d operator-(Matrix3d&);
-	float& operator()(int i,int j){return mat[i][j];}
+	double& operator()(int i,int j){return mat[i][j];}
 
 	Matrix3d inverse();
-	static Matrix3d skew(vector3<float>& v);
+	static Matrix3d skew(vector3<double>& v);
 
 	~Matrix3d(void);
 };
 
 inline
-float Matrix3d::determinant()
+double Matrix3d::determinant()
 {
 	return mat[0][0] * (mat[1][1] * mat[2][2] - mat[1][2] * mat[2][1]) 
 		 - mat[0][1] * (mat[2][2] * mat[1][0] - mat[1][2] * mat[2][0]) 
@@ -42,7 +42,7 @@ return Matrix3d(mat[0][0],mat[1][0],mat[2][0],
 
 inline
 Matrix3d Matrix3d::operator*(Matrix3d& mm){
-	float m[3][3]={0};
+	double m[3][3]={0};
 	for(int i=0;i<3;i++)
 		for(int j=0;j<3;j++)
 			for(int k=0;k<3;k++)
@@ -51,16 +51,16 @@ Matrix3d Matrix3d::operator*(Matrix3d& mm){
 }
 
 inline
-vector3<float> Matrix3d::operator*(vector3<float>& vv){
-	float v[3]={0};
+vector3<double> Matrix3d::operator*(vector3<double>& vv){
+	double v[3]={0};
 	for(int i=0;i<3;i++)
 		for(int j=0;j<3;j++)
 			v[i]+=mat[i][j] * vv.coords[j];
-	return vector3<float>(v);
+	return vector3<double>(v);
 }
 
 inline
-Matrix3d Matrix3d::operator*(float k){
+Matrix3d Matrix3d::operator*(double k){
 	for(int i=0;i<3;i++)
 		for(int j=0;j<3;j++)
 				mat[i][j] *= k;
@@ -69,7 +69,7 @@ Matrix3d Matrix3d::operator*(float k){
 
 inline
 Matrix3d Matrix3d::operator+(Matrix3d& mm){
-	float m[3][3];
+	double m[3][3];
 	for(int i=0;i<3;i++)
 		for(int j=0;j<3;j++)
 			m[i][j] = mat[i][j]+mm.mat[i][j];
@@ -78,7 +78,7 @@ Matrix3d Matrix3d::operator+(Matrix3d& mm){
 
 inline
 Matrix3d Matrix3d::operator-(Matrix3d& mm){
-	float m[3][3];
+	double m[3][3];
 	for(int i=0;i<3;i++)
 		for(int j=0;j<3;j++)
 			m[i][j] = mat[i][j]-mm.mat[i][j];
